add "data header" + CG TODO: print answer

+++ potapova.marina/I6/calculateAtanSum.cpp
@@ -11,7 +11,7 @@ double calculateAtanSum(double x, double abs_error, size_t num_max)
     double numerator = x;
     double summation = x;
     int sign = 1;
-    int num_summation = 1;
+    size_t num_summation = 1;
     double sum_res = 0;
     while (fabs(summation) > abs_error)
     {
